+++ src/lib/fetch.ts
@@ -19,6 +19,7 @@ import {
 export {
   fetchAccount,
   fetchLastBlock,
+  checkZkappTransaction,
   parseFetchedAccount,
   markAccountToBeFetched,
   markNetworkToBeFetched,
@@ -135,7 +136,7 @@ async function fetchAccountInternal(
 }
 
 type FetchConfig = { timeout?: number };
-type FetchResponse = { data: any };
+type FetchResponse = { data: any; errors?: any };
 type FetchError = {
   statusCode: number;
   statusText: string;
@@ -389,6 +390,84 @@ const lastBlockQuery = `{
     }
   }
 }`;
+
+type LastBlockQueryFailureCheckResponse = {
+  bestChain: {
+    transactions: {
+      zkappCommands: {
+        hash: string;
+        failureReason: {
+          failures: string[];
+          index: number;
+        }[];
+      }[];
+    };
+  }[];
+};
+
+const lastBlockQueryFailureCheck = `{
+  bestChain(maxLength: 1) {
+    transactions {
+      zkappCommands {
+        hash
+        failureReason {
+          failures
+          index
+        }
+      }
+    }
+  }
+}`;
+
+async function fetchLatestBlockZkappStatus(
+  graphqlEndpoint = defaultGraphqlEndpoint
+) {
+  let [resp, error] = await makeGraphqlRequest(
+    lastBlockQueryFailureCheck,
+    graphqlEndpoint
+  );
+  if (error) throw Error(`Error making GraphQL request: ${error.statusText}`);
+  let bestChain = resp?.data as LastBlockQueryFailureCheckResponse;
+  if (bestChain === undefined) {
+    throw Error(
+      'Failed to fetch the latest zkApp transaction status. The response data is undefined.'
+    );
+  }
+  return bestChain;
+}
+
+async function checkZkappTransaction(txnId: string) {
+  let bestChainBlocks = await fetchLatestBlockZkappStatus();
+
+  for (let block of bestChainBlocks.bestChain) {
+    for (let zkappCommand of block.transactions.zkappCommands) {
+      if (zkappCommand.hash === txnId) {
+        if (zkappCommand.failureReason !== null) {
+          let failureReason = zkappCommand.failureReason
+            .reverse()
+            .map((failure) => {
+              return ` AccountUpdate #${
+                failure.index
+              } failed. Reason: "${failure.failures.join(', ')}"`;
+            });
+          return {
+            success: false,
+            failureReason,
+          };
+        } else {
+          return {
+            success: true,
+            failureReason: null,
+          };
+        }
+      }
+    }
+  }
+  return {
+    success: false,
+    failureReason: null,
+  };
+}
 
 type FetchedBlock = {
   protocolState: {
@@ -892,7 +971,17 @@ async function checkResponseStatus(
   response: Response
 ): Promise<[FetchResponse, undefined] | [undefined, FetchError]> {
   if (response.ok) {
-    return [(await response.json()) as FetchResponse, undefined];
+    let jsonResponse = await response.json();
+    if (jsonResponse.errors && jsonResponse.errors.length > 0) {
+      return [
+        undefined,
+        {
+          statusCode: response.status,
+          statusText: jsonResponse.errors,
+        } as FetchError,
+      ];
+    }
+    return [jsonResponse as FetchResponse, undefined];
   } else {
     return [
       undefined,